+++ tests/test_functions/test_data/test_data.py
@@ -1,31 +1,31 @@
 from dublib.functions import data
 
-def test_Copy():
+def test_deep_copy():
 	Dictionary = {"test": [1]}
-	Copy = data.Copy(Dictionary)
+	Copy = data.deep_copy(Dictionary)
 	assert Copy is not Dictionary
 	assert Copy["test"] is not Dictionary["test"]
 
-def test_StringifyFloat():
-	assert data.StringifyFloat(1.056) == "1.06"
-	assert data.StringifyFloat(1.054) == "1.05"
-	assert data.StringifyFloat(1.0) == "1"
-	assert data.StringifyFloat(0.00) == "0"
+def test_stringify_float():
+	assert data.stringify_float(1.056) == "1.06"
+	assert data.stringify_float(1.054) == "1.05"
+	assert data.stringify_float(1.0) == "1"
+	assert data.stringify_float(0.00) == "0"
 
-def test_StringToBool():
-	assert data.StringToBool("0") is False
-	assert data.StringToBool("fAlsE") is False
-	assert data.StringToBool("1") is True
-	assert data.StringToBool("") is False
+def test_string_to_bool():
+	assert data.string_to_bool("0") is False
+	assert data.string_to_bool("fAlsE") is False
+	assert data.string_to_bool("1") is True
+	assert data.string_to_bool("") is False
 
 
-def test_ToSequence():
-	assert data.ToSequence("test") == ("test",)
-	assert data.ToSequence(["test"]) == ("test",)
-	assert data.ToSequence("test", target_type = list) == ["test"]
+def test_to_sequence():
+	assert data.to_sequence("test") == ("test",)
+	assert data.to_sequence(["test"]) == ("test",)
+	assert data.to_sequence("test", target_type = list) == ["test"]
 
-def test_Zerotify():
-	assert data.Zerotify(0) is None
-	assert data.Zerotify("") is None
-	assert data.Zerotify(-1) == -1
-	assert data.Zerotify("1") == "1"
+def test_zerotify():
+	assert data.zerotify(0) is None
+	assert data.zerotify("") is None
+	assert data.zerotify(-1) == -1
+	assert data.zerotify("1") == "1"